Ajout de l'authentification dans le controller

+++ Automation/Controller/AutomationController.cs
@@ -1,5 +1,6 @@
 ﻿using Automation.Dto.Request;
 using Automation.Service;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Automation.Controller;
@@ -7,8 +8,10 @@ namespace Automation.Controller;
 [ApiController]
 [Route("/api/[controller]")]
 [Produces("application/json")]
+[Authorize]
 public class AutomationController : ControllerBase
 {
+    
     private readonly AutomationService _automationService;
 
     public AutomationController(AutomationService automationService)
@@ -17,14 +20,25 @@ public class AutomationController : ControllerBase
     }
 
     [HttpGet]
-    public IActionResult GetAutomations(string username)
+    
+    public IActionResult GetAutomations()
     {
+        var username = User.Claims.ElementAtOrDefault(1)?.Value;
+        if (username == null)
+        {
+            return Unauthorized();
+        }
         return Ok(_automationService.GetAutomations(username));
     }
 
     [HttpDelete]
-    public IActionResult SupprimerAutomation(int idAutomation, string username)
+    public IActionResult SupprimerAutomation(int idAutomation)
     {
+        var username = User.Claims.ElementAtOrDefault(1)?.Value;
+        if (username == null)
+        {
+            return Unauthorized();
+        }
         _automationService.DeleteAutomation(idAutomation, username);
         return NoContent();
     }
@@ -32,15 +46,25 @@ public class AutomationController : ControllerBase
     [HttpPost("/api/dca")]
     public IActionResult AjouterDca([FromBody] DcaReqDto req)
     {
-        _automationService.AjouterDca(req.Username, req.Symbole, req.Quantite, req.Frequence);
+        var username = User.Claims.ElementAtOrDefault(1)?.Value;
+        if (username == null || username!=req.Username)
+        {
+            return Unauthorized();
+        }
+        _automationService.AjouterDca(username, req.Symbole, req.Quantite, req.Frequence);
         return Ok();
     }
 
     [HttpPost("/api/pricethreshold")]
     public IActionResult AjouterPriceThreshold([FromBody] PriceThresholdReqDto req)
     {
+        var username = User.Claims.ElementAtOrDefault(1)?.Value;
+        if (username == null || username!=req.Username)
+        {
+            return Unauthorized();
+        }
         _automationService.AjouterPriceThreshold(req.ThresholdPrice, req.TransactionType, req.ThresholdType,
-            req.Username);
+            username);
         return Ok();
     }
 }
